support matching of different tokens via regexes
This patch implements a match registry that can provide a set of
different regular expressions to be registered in the terminal.

This not only allows us to perform correct matching of more complex URLs
but also file links, email addresses, VoIP links, man pages and numbers.

LICENSE NOTES

This includes code taken from gnome-terminal, which is licensed under
GPLv3. According to the [compatibility matrix][1] of the GNU GPL license
FAQ this means that existing code in tilda can remain "GPLv2 or later"
licensed, but the combined code is now licensed under "GPLv3 or later".

Fixes: #405

[1]: https://www.gnu.org/licenses/gpl-faq.html#AllCompatibility